pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "codagent"
version = "0.1.0"
description = "Terminal-based AI coding agent"
requires-python = ">=3.10"
dependencies = [
    "pydantic-ai-slim[openai,anthropic,openrouter]>=1.63.0",
    "python-dotenv>=1.2.1",
]

[project.scripts]
codagent = "codagent.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/codagent"]
